Move neural nets factory out of utils (#994)
* Export classes and function

* Reworked factory imports

* Debugging ruff

* Debugging ruff

* Added links to the new location with deprecation warnings

* Forgot build_zuko_maf

* Added tests for the deprecated functions

* Renaming

* Doc strings now only contain the deprecation warning

* Updated the deprecation message

* Updated the deprecation message

* Ruff once more
