+++ sbi/inference/snpe/snpe_c.py
@@ -51,9 +51,9 @@ class SNPE_C(PosteriorEstimator):
         In this codebase, we will automatically switch to the non-atomic loss if the
         following criteria are fulfilled:<br/>
         - proposal is a `DirectPosterior` with density_estimator `mdn`, as built
-            with `utils.sbi.posterior_nn()`.<br/>
+            with `sbi.neural_nets.posterior_nn()`.<br/>
         - the density estimator is a `mdn`, as built with
-            `utils.sbi.posterior_nn()`.<br/>
+            `sbi.neural_nets.posterior_nn()`.<br/>
         - `isinstance(prior, MultivariateNormal)` (from `torch.distributions`) or
             `isinstance(prior, sbi.utils.BoxUniform)`
