+++ sbi/neural_nets/__init__.py
@@ -1,0 +1,23 @@
+from sbi.neural_nets.classifier import (
+    StandardizeInputs,
+    build_input_layer,
+    build_linear_classifier,
+    build_mlp_classifier,
+    build_resnet_classifier,
+)
+from sbi.neural_nets.density_estimators import DensityEstimator, NFlowsFlow
+from sbi.neural_nets.embedding_nets import (
+    CNNEmbedding,
+    FCEmbedding,
+    PermutationInvariantEmbedding,
+)
+from sbi.neural_nets.factory import classifier_nn, likelihood_nn, posterior_nn
+from sbi.neural_nets.flow import (
+    build_made,
+    build_maf,
+    build_maf_rqs,
+    build_nsf,
+    build_zuko_maf,
+)
+from sbi.neural_nets.mdn import build_mdn
+from sbi.neural_nets.mnle import MixedDensityEstimator, build_mnle
