+++ sbi/utils/__init__.py
@@ -1,7 +1,6 @@
 # flake8: noqa
 from sbi.utils.analysis_utils import get_1d_marginal_peaks_from_kde
 from sbi.utils.conditional_density_utils import extract_and_transform_mog
-from sbi.utils.get_nn_models import classifier_nn, likelihood_nn, posterior_nn
 from sbi.utils.io import get_data_root, get_log_root, get_project_root
 from sbi.utils.kde import KDEWrapper, get_kde
 from sbi.utils.potentialutils import pyro_potential_wrapper, transformed_potential
