+++ tests/inference_on_device_test.py
@@ -29,8 +29,8 @@ from sbi.inference import (
 )
 from sbi.inference.posteriors.importance_posterior import ImportanceSamplingPosterior
 from sbi.inference.potentials.base_potential import BasePotential
+from sbi.neural_nets import classifier_nn, likelihood_nn, posterior_nn
 from sbi.simulators import diagonal_linear_gaussian, linear_gaussian
-from sbi.utils.get_nn_models import classifier_nn, likelihood_nn, posterior_nn
 from sbi.utils.torchutils import BoxUniform, gpu_available, process_device
 from sbi.utils.user_input_checks import (
     check_embedding_net_device,
@@ -110,20 +110,20 @@ def test_training_and_mcmc_on_device(
 
     if method in [SNPE_A, SNPE_C]:
         kwargs = dict(
-            density_estimator=utils.posterior_nn(
+            density_estimator=posterior_nn(
                 model=model, num_transforms=2, dtype=torch.float32
             )
         )
         train_kwargs = dict(force_first_round_loss=True)
     elif method == SNLE:
         kwargs = dict(
-            density_estimator=utils.likelihood_nn(
+            density_estimator=likelihood_nn(
                 model=model, num_transforms=2, dtype=torch.float32
             )
         )
         train_kwargs = dict()
     elif method in (SNRE_A, SNRE_B, SNRE_C):
-        kwargs = dict(classifier=utils.classifier_nn(model=model))
+        kwargs = dict(classifier=classifier_nn(model=model))
         train_kwargs = dict()
     else:
         raise ValueError()
