+++ tests/linearGaussian_snle_test.py
@@ -18,6 +18,7 @@ from sbi.inference import (
     prepare_for_sbi,
     simulate_for_sbi,
 )
+from sbi.neural_nets import likelihood_nn
 from sbi.simulators.linear_gaussian import (
     diagonal_linear_gaussian,
     linear_gaussian,
@@ -25,7 +26,7 @@ from sbi.simulators.linear_gaussian import (
     samples_true_posterior_linear_gaussian_uniform_prior,
     true_posterior_linear_gaussian_mvn_prior,
 )
-from sbi.utils import BoxUniform, likelihood_nn, process_prior
+from sbi.utils import BoxUniform, process_prior
 
 from .test_utils import check_c2st, get_prob_outside_uniform_prior
